Dumbify identicon
