+++ src/cow-react/modules/account/containers/AccountDetails/index.tsx
@@ -11,7 +11,7 @@ import { getConnection } from '@cow/modules/wallet/web3-react/utils/connection'
 import CoinbaseWalletIcon from '@cow/modules/wallet/api/assets/coinbase.svg'
 import WalletConnectIcon from '@cow/modules/wallet/api/assets/walletConnectIcon.svg'
 import FortmaticIcon from '@cow/modules/wallet/api/assets/formatic.png'
-import { Identicon } from '@cow/modules/wallet/web3-react/containers/Identicon'
+import { Identicon } from '@cow/modules/wallet/api/container/Identicon'
 import { ActivityDescriptors } from 'hooks/useRecentActivity'
 import Activity from '@cow/modules/account/containers/Transaction'
