+++ src/cow-react/modules/wallet/index.ts
@@ -5,5 +5,5 @@ export * from './web3-react/hooks/useDisconnectWallet'
 
 export * from './web3-react/containers/Web3Status'
 export * from './web3-react/containers/WalletModal'
-export * from './web3-react/containers/Identicon'
+export * from './api/container/Identicon'
 export * from './web3-react/containers/AddToMetamask'
